test_cases/requirements.txt
pytest>=8.0
requests>=2.31
python-socketio[client]>=5.11.0
